game renderer moved #207
