+++ frontend/src/models/Services/game_renderer.ts
